move exceptions to their own folder and namespace; handle 'MyFileNotFoundException' in app logic; refactor fields and constants to Pascal Case; format and arrange code to be easier to read.
