+++ CredentialsManager/Credentials.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using CredentialsManager.FilesExceptions;
 
 namespace CredentialsManager
 {
